Practice exercise
Did practice for D.O.M - practice adding a new element in the Linkedin course.

There are plenty of ways of doing this. I could have added the HTML as a bit of content inside an object, and done the css inside the HTML as <style>. Watch the solution video to see how the dude did it